let's merge layers!
